move typecheck formatter to program_processors.formatters
